+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `GetTheatre` API
 - `GetUnitType` API
 
+### Changed
+- Improved reliability of error handling.
+
 ## [0.5.0] - 2022-04-19
 ### Added
 - `GetMissionFilename` API